Small refactor: / 2 --> / 2d

+++ Graphics/src/main/java/dk/lima/graphics/gameRender/EntityRenderer.java
@@ -58,8 +58,8 @@ public class EntityRenderer implements IGraphicsService {
 
             TransformCP transformCP = entity.getComponent(TransformCP.class);
             if (entity instanceof Player) {
-                polygon.setTranslateX((double) gameData.getDisplayWidth() / 2);
-                polygon.setTranslateY((double) gameData.getDisplayHeight() / 2);
+                polygon.setTranslateX(gameData.getDisplayWidth() / 2d);
+                polygon.setTranslateY(gameData.getDisplayHeight() / 2d);
             } else {
                 Coordinate coord = transformCP.getCoord();
                 polygon.setTranslateX(coord.getX() + gameData.getDisplayWidth() / 2d - world.getPlayerPosition().getX());
